chore(recipe): move recipe persistance schema to recipe.schema
